requirements.txt
Django==3.1.7
decorator==5.0.5
bootstrap-admin==0.4.*
django-debug-toolbar==3.2
python-dateutil==2.8.1
sqlparse==0.4.1
sphinx
sphinx_rtd_theme
django-ratelimit==1.1.0
coverage
django-crispy-forms==1.11.2
channels-redis
daphne
psycopg2
dj-database-url
django-redis
redis
huey
sentry-sdk==1.0.0
crispy-bootstrap5
asgiref==3.2.10
